refactor: extract IngestionSnapshotLifecycle and remove dead RekeyTenantSoftwareReferences
